requirements.txt
# Core BigQuery and data libraries
google-cloud-bigquery>=3.4.0
db-dtypes>=1.0.0
pandas>=1.5.0
numpy>=1.23.0

# Configuration
pyyaml>=6.0

# Analysis libraries
networkx>=3.0
scipy>=1.9.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.16.0